+++ erddapy/core/url.py
@@ -2,10 +2,14 @@
 
 import functools
 import io
-from typing import Dict, Optional
+from datetime import datetime
+from typing import Dict, Optional, Union
 from typing.io import BinaryIO
+from urllib.parse import quote_plus
 
 import httpx
+import pytz
+from pandas._libs.tslibs.parsing import parse_time_string
 
 
 @functools.lru_cache(maxsize=256)
@@ -76,3 +80,140 @@ def _multi_urlopen(url: str) -> BinaryIO:
     except (httpx.HTTPError, httpx.ConnectError):
         return None
     return data
+
+
+def _quote_string_constraints(kwargs: Dict) -> Dict:
+    """
+    Quote constraints of String variables.
+
+    The right-hand-side value must be surrounded by double quotes if they are not relative constraints.
+    """
+    return {
+        k: f'"{v}"' if isinstance(v, str) and not _check_substrings(v) else v
+        for k, v in kwargs.items()
+    }
+
+
+def _format_constraints_url(kwargs: Dict) -> str:
+    """Join the constraint variables with separator '&' to add to the download link."""
+    return "".join([f"&{k}{v}" for k, v in kwargs.items()])
+
+
+def _check_substrings(constraint):
+    """Extend the OPeNDAP with extra strings."""
+    substrings = ["now", "min", "max"]
+    return any([True for substring in substrings if substring in str(constraint)])
+
+
+def parse_dates(date_time: Union[datetime, str]) -> float:
+    """
+    Parse dates to ERDDAP internal format.
+
+    ERDDAP ReSTful API standardizes the representation of dates as either ISO
+    strings or seconds since 1970, but internally ERDDAPY uses datetime-like
+    objects. `timestamp` returns the expected strings in seconds since 1970.
+
+    """
+    if isinstance(date_time, str):
+        # pandas returns a tuple with datetime, dateutil, and string representation.
+        # we want only the datetime obj.
+        parse_date_time = parse_time_string(date_time)[0]
+    else:
+        parse_date_time = date_time
+
+    if not parse_date_time.tzinfo:
+        parse_date_time = pytz.utc.localize(parse_date_time)
+    else:
+        parse_date_time = parse_date_time.astimezone(pytz.utc)
+
+    return parse_date_time.timestamp()
+
+
+def _search_url(
+    server: str,
+    response: str = "html",
+    search_for: Optional[str] = None,
+    protocol: str = "tabledap",
+    items_per_page: int = 1000,
+    page: int = 1,
+    **kwargs,
+):
+    server = server.rstrip("/")
+    base = (
+        "{server}/search/advanced.{response}"
+        "?page={page}"
+        "&itemsPerPage={itemsPerPage}"
+        "&protocol={protocol}"
+        "&cdm_data_type={cdm_data_type}"
+        "&institution={institution}"
+        "&ioos_category={ioos_category}"
+        "&keywords={keywords}"
+        "&long_name={long_name}"
+        "&standard_name={standard_name}"
+        "&variableName={variableName}"
+        "&minLon={minLon}"
+        "&maxLon={maxLon}"
+        "&minLat={minLat}"
+        "&maxLat={maxLat}"
+        "&minTime={minTime}"
+        "&maxTime={maxTime}"
+    )
+    if search_for:
+        search_for = quote_plus(search_for)
+        base += "&searchFor={searchFor}"
+
+    # Convert dates from datetime to `seconds since 1970-01-01T00:00:00Z`.
+    min_time = kwargs.pop("min_time", "")
+    max_time = kwargs.pop("max_time", "")
+    if min_time and not _check_substrings(min_time):
+        kwargs.update({"min_time": parse_dates(min_time)})
+    else:
+        kwargs.update({"min_time": min_time})
+    if max_time and not _check_substrings(max_time):
+        kwargs.update({"max_time": parse_dates(max_time)})
+    else:
+        kwargs.update({"max_time": max_time})
+
+    if protocol:
+        kwargs.update({"protocol": protocol})
+
+    lower_case_search_terms = (
+        "cdm_data_type",
+        "institution",
+        "ioos_category",
+        "keywords",
+        "long_name",
+        "standard_name",
+        "variableName",
+    )
+    for search_term in lower_case_search_terms:
+        if search_term in kwargs.keys():
+            lowercase = kwargs[search_term].lower()
+            kwargs.update({search_term: lowercase})
+
+    default = "(ANY)"
+    url = base.format(
+        server=server,
+        response=response,
+        page=page,
+        itemsPerPage=items_per_page,
+        protocol=kwargs.get("protocol", default),
+        cdm_data_type=kwargs.get("cdm_data_type", default),
+        institution=kwargs.get("institution", default),
+        ioos_category=kwargs.get("ioos_category", default),
+        keywords=kwargs.get("keywords", default),
+        long_name=kwargs.get("long_name", default),
+        standard_name=kwargs.get("standard_name", default),
+        variableName=kwargs.get("variableName", default),
+        minLon=kwargs.get("min_lon", default),
+        maxLon=kwargs.get("max_lon", default),
+        minLat=kwargs.get("min_lat", default),
+        maxLat=kwargs.get("max_lat", default),
+        minTime=kwargs.get("min_time", default),
+        maxTime=kwargs.get("max_time", default),
+        searchFor=search_for,
+    )
+    # ERDDAP 2.10 no longer accepts strings placeholder for dates.
+    # Removing them entirely should be OK for older versions too.
+    url = url.replace("&minTime=(ANY)", "").replace("&maxTime=(ANY)", "")
+    return url
