+++ erddapy/multiple_server_search.py
@@ -11,8 +11,7 @@ try:
 except ImportError:
     joblib = False
 
-from erddapy.core.url import _format_search_string, _multi_urlopen
-from erddapy.erddapy import _search_url
+from erddapy.core.url import _format_search_string, _multi_urlopen, _search_url
 from erddapy.servers.servers import servers
